Add first working version of the Sudoku solving script
IMPORTANT: this is not yet finished. If you have two remaining empty cells
in one row/column and you enter a number into one of them, the other one will
be filled in as well, but it won't be removed as a hint from surrounding cells.
Other solving strategies will be added soon; code readability will improve.